#### **Overview**  
This PR implements preliminary support for Minecraft Bedrock version
**1.21.70** ahead of its official release. The goal is to minimize the
update time once the version is officially available by preparing most
of the necessary changes in advance. Once 1.21.70 is released, only
function offsets will need to be updated, ensuring a seamless transition
with minimal disruption to server owners.

- This PR has been tested on the latest **1.21.70 preview build
(beta.26)** on both **Windows** and **Linux**.
- Server owners can use the preview server support as a temporary
solution when 1.21.70 is officially released, allowing players with the
new client to connect before full support is completed.
- Once **1.21.70** is officially released, we will update the offsets as
soon as possible. Since most of the work has already been completed in
this PR, full 1.21.70 support should be finalized within a day.

+++ src/bedrock/util/timer.h
@@ -48,7 +48,7 @@ private:
     int64_t last_ms_;
     int64_t last_ms_sys_time_;
     float adjust_time_;
-    int stepping_tick_;
+    float stepping_tick_;
     std::function<std::int64_t()> get_time_ms_callback_;
     static constexpr int MAX_TICKS_PER_UPDATE = 10;
     static constexpr float MAX_TIME_PER_UPDATE = 1.0 / MAX_TICKS_PER_UPDATE;
